# Sistema_Crud_API_PIBIC/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SistemaAPI PIBIC
  version: 1.0.0
  description: |
    API REST para gerenciamento e resolucao de casos clinicos academicos.
    As operacoes de IA aceitam somente dados sinteticos ou previamente
    desidentificados e toda saida exige revisao humana antes da publicacao.
    O recurso e educacional e nao substitui decisao clinica.
  contact:
    name: Equipe SistemaAPI PIBIC
  license:
    name: Uso academico
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /alunos/{id}/casos/{casoId}/completo:
    get:
      summary: Abre o caso publicado e inicia ou retoma a tentativa do aluno
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/casoId"
      responses:
        "200":
          description: Caso completo publicado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoAluno"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
...
  /perguntas/caso/{casoId}:
    get:
      summary: Lista perguntas de um caso com paginacao
      parameters:
        - $ref: "#/components/parameters/casoId"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de perguntas do caso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaPergunta"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    post:
      summary: Cria pergunta no caso indicado no caminho
      description: O caso do caminho e usado; idCaso no corpo e opcional.
      parameters:
        - $ref: "#/components/parameters/casoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerguntaRequest"
      responses:
        "201":
          description: Pergunta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerguntaResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /casos:
    get:
      summary: Lista casos clinicos com paginacao e filtros
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [RASCUNHO, PUBLICADO, ARQUIVADO]
        - name: idProfessor
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: termo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            example: dataCriacao,desc
      responses:
        "200":
          description: Casos encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaCasoClinico"
    post:
      summary: Cria caso clinico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasoClinicoRequest"
      responses:
        "201":
          description: Caso criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoResponse"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /casos/{id}:
    get:
      summary: Busca caso por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Caso encontrado
    put:
      summary: Atualiza caso
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasoClinicoRequest"
      responses:
        "200":
          description: Caso atualizado
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
    delete:
      summary: Remove caso
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Caso removido
        "409":
          $ref: "#/components/responses/Conflito"
  /casos/{id}/completo:
    get:
      summary: Busca caso completo com pacientes, conteudos e perguntas
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Caso completo
  /casos/{id}/publicar:
    patch:
      summary: Publica caso clinico e aplica sua configuracao atomica
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: false
        description: Configuracao opcional aplicada na mesma transacao da publicacao.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicarCasoRequest"
      responses:
        "200":
          description: Caso publicado
        "409":
          $ref: "#/components/responses/Conflito"
  /casos/{id}/arquivar:
    patch:
      summary: Arquiva um caso clinico publicado
      description: |
        Remove o caso do catalogo dos alunos sem apagar seu historico. A
        operacao e idempotente quando o caso ja esta arquivado.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Caso arquivado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoResponse"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
...
  /casos/{id}/respostas/pendentes-revisao:
    get:
      summary: Lista respostas discursivas ou de conduta pendentes de revisao
      description: Disponivel ao professor responsavel pelo caso e a administradores.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          description: Ordenacao no formato campo,direcao.
          schema:
            type: string
            default: dataResposta,asc
      responses:
        "200":
          description: Pagina de respostas que aguardam revisao humana
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaRespostaAluno"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /casos/{id}/respostas/{idResposta}/revisao:
    patch:
      summary: Registra a revisao humana de uma resposta discursiva ou de conduta
      description: |
        A mesma decisao pode ser reenviada de forma idempotente. Uma decisao
        diferente para uma resposta ja revisada retorna conflito.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: idResposta
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisarRespostaRequestDTO"
      responses:
        "200":
          description: Resposta revisada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisaoResposta"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /casos/{id}/respostas/{idResposta}/revisoes:
    get:
      summary: Lista o historico imutavel de revisoes de uma resposta
      parameters:
        - $ref: "#/components/parameters/id"
        - name: idResposta
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Revisoes em ordem cronologica
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevisaoResposta"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /casos/{id}/respostas/{idResposta}/revisoes/pagina:
    get:
      summary: Lista de forma paginada o historico imutavel de revisoes de uma resposta
      parameters:
        - $ref: "#/components/parameters/id"
        - name: idResposta
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de revisoes em ordem cronologica
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaRevisaoResposta"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /casos/{casoId}/perguntas:
    get:
      summary: Lista perguntas de um caso
      parameters:
        - $ref: "#/components/parameters/casoId"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Perguntas encontradas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaPergunta"
    post:
      summary: Cria pergunta vinculada ao caso
      parameters:
        - $ref: "#/components/parameters/casoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerguntaRequest"
      responses:
        "201":
          description: Pergunta criada
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    casoId:
      name: casoId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    page:
      name: page
      in: query
      description: Indice da pagina, iniciando em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: Quantidade de elementos por pagina.
      schema:
        type: integer
        minimum: 1
        default: 20
    sort:
      name: sort
      in: query
      description: Ordenacao no formato campo,direcao.
      schema:
        type: string
        example: nome,asc
  responses:
    CorpoRequisicaoExcedido:
      description: O corpo da requisicao excede o limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:corpo-requisicao-excedido
            title: Corpo da requisicao excedido
            status: 413
            detail: O corpo da requisicao excede o limite permitido
            instance: /auth/login
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O corpo da requisicao excede o limite permitido
...
    DadosInvalidos:
      description: A requisicao possui campos, JSON, parametros ou tipos invalidos.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    AcessoNegado:
      description: O usuario autenticado nao possui permissao para a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NaoEncontrado:
      description: O recurso solicitado nao foi encontrado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflito:
      description: O estado atual ou a concorrencia dos dados impede a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos ate a repeticao recomendada quando o conflito for transitorio.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    CorrelationId:
      description: Identificador da requisicao para rastreamento ponta a ponta.
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
    PublicarCasoRequest:
      type: object
      required:
        - tempoLimiteMinutos
      properties:
        tempoLimiteMinutos:
          type: integer
          minimum: 15
          maximum: 480
          example: 60
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - correlationId
        - erro
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        erro:
          type: string
          deprecated: true
          description: Alias legado de detail; mantido temporariamente para compatibilidade.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo quando a validacao falha.
        campos:
          type: object
          deprecated: true
          additionalProperties:
            type: string
          description: Alias legado de errors; mantido temporariamente para compatibilidade.
...
    CasoClinicoRequest:
      type: object
      required:
        - titulo
        - disciplina
        - areaSaude
        - estilo
        - especialidade
        - nivelDificuldade
      properties:
        idProfessor:
          type: integer
          format: int64
          nullable: true
        titulo:
          type: string
          maxLength: 200
        disciplina:
          type: string
          maxLength: 120
        areaSaude:
          type: string
          maxLength: 120
        estilo:
          type: string
          maxLength: 60
        especialidade:
          type: string
          maxLength: 120
        objetivoAprendizagem:
          type: string
          maxLength: 10000
        nivelDificuldade:
          type: string
          enum: [BAIXA, MEDIA, ALTA]
        tempoLimiteMinutos:
          type: integer
          minimum: 15
          maximum: 480
          default: 60
          description: Tempo total para responder o caso completo.
    CasoClinicoResponse:
      allOf:
        - $ref: "#/components/schemas/CasoClinicoRequest"
        - type: object
          properties:
            idCaso:
              type: integer
              format: int64
            status:
              type: string
              enum: [RASCUNHO, PUBLICADO, ARQUIVADO]
            dataCriacao:
              type: string
              format: date-time
            dataAtualizacao:
              type: string
              format: date-time
...
    AlternativaPergunta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        letra:
          type: string
          maxLength: 10
        texto:
          type: string
          maxLength: 10000
        correta:
          type: boolean
    AlternativaAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        letra:
          type: string
          maxLength: 10
        texto:
          type: string
          maxLength: 10000
    PerguntaRequest:
      type: object
      required:
        - texto
        - tipo
        - gabarito
        - resposta
      properties:
        idCaso:
          type: integer
          format: int64
          nullable: true
        texto:
          type: string
          maxLength: 10000
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
          example: MULTIPLA_ESCOLHA
        gabarito:
          type: string
          maxLength: 10000
        resposta:
          type: string
          maxLength: 10000
        alternativas:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/AlternativaPergunta"
    PerguntaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        texto:
          type: string
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
          example: MULTIPLA_ESCOLHA
        gabarito:
          type: string
        resposta:
          type: string
        alternativas:
          type: array
          items:
            $ref: "#/components/schemas/AlternativaPergunta"
    PerguntaAluno:
      type: object
      description: Pergunta sem campos de correcao ou gabarito.
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        texto:
          type: string
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
        alternativas:
          type: array
          items:
            $ref: "#/components/schemas/AlternativaAluno"
    ConteudoClinicoAluno:
      type: object
      description: Conteudo clinico sem o diagnostico esperado.
      properties:
        idConteudo:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        sintomas:
          type: string
        contexto:
          type: string
        examClinico:
          type: string
        antecClinico:
          type: string
    CasoClinicoAluno:
      type: object
      description: Visao segura do caso destinada ao aluno.
      properties:
        caso:
          $ref: "#/components/schemas/CasoClinicoResponse"
        pacientes:
          type: array
          items:
            type: object
        conteudosClinicos:
          type: array
          items:
            $ref: "#/components/schemas/ConteudoClinicoAluno"
        perguntas:
          type: array
          items:
            $ref: "#/components/schemas/PerguntaAluno"
        inicioTentativa:
          type: string
          format: date-time
        prazoFinal:
          type: string
          format: date-time
        segundosRestantes:
          type: integer
          format: int64
    RespostaAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idAluno:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        idPergunta:
          type: integer
          format: int64
        respostaMarcada:
          type: string
        correta:
          type: boolean
          nullable: true
          description: |
            Resultado da correcao automatica ou da revisao humana. O valor
            null indica que a resposta ainda aguarda revisao.
        dataResposta:
          type: string
          format: date-time
    RevisarRespostaRequestDTO:
      type: object
      required:
        - correta
        - justificativa
      properties:
        correta:
          type: boolean
          description: Decisao do professor sobre a correcao da resposta.
        justificativa:
          type: string
          minLength: 1
          maxLength: 2000
          description: Fundamentacao obrigatoria e auditavel da decisao humana.
    RevisaoResposta:
      type: object
      required:
        - id
        - idResposta
        - idRevisor
        - correta
        - justificativa
        - dataRevisao
        - versaoRevisao
      properties:
        id:
          type: integer
          format: int64
        idResposta:
          type: integer
          format: int64
        idRevisor:
          type: integer
          format: int64
        correta:
          type: boolean
        justificativa:
          type: string
          maxLength: 2000
        dataRevisao:
          type: string
          format: date-time
        versaoRevisao:
          type: integer
          format: int64
          minimum: 1
    PaginaBase:
      type: object
      required:
        - totalElements
        - totalPages
        - number
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaginaCasoClinico:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/CasoClinicoResponse"
...
    PaginaPergunta:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/PerguntaResponse"
...
    PaginaRespostaAluno:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/RespostaAluno"
    PaginaRevisaoResposta:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/RevisaoResposta"
...